+++ lib/sources/ads.class.php
@@ -313,8 +313,8 @@ class _ads {
 		if (isset($_POST['orders']))
 			$orders = (array)$_POST['orders'];
 		
-		if (isset($_GET['delete']))
-			$delete = (int)$_GET['delete'];
+		if (isset($_POST['delete']))
+			$delete = (int)$_POST['delete'];
 		
 		if (isset($_GET['edit']))
 			$edit = (int)$_GET['edit'];
@@ -351,6 +351,12 @@ class _ads {
 		}
 		
 		if ($delete) {
+			if (!security::checkToken()) {
+				api::callHooks(API_HOOK_AFTER,
+					'ads::verifyAdmin', $this, $form);
+				return false;
+			}
+			
 			$result = $this->delete($id);
 			
 			if ($result)
@@ -840,6 +846,18 @@ class _ads {
 		echo
 			"<div class='admin-content'>";
 				
+		if ($delete && $id && empty($_POST['delete'])) {
+			$selected = sql::fetch(sql::run(
+				" SELECT `Title` FROM `{ads}`" .
+				" WHERE `ID` = '".$id."'" .
+				($this->userPermissionIDs?
+					" AND `ID` IN (".$this->userPermissionIDs.")":
+					null)));
+			
+			security::displayConfirmation(
+				'<b>'.__('Delete').'?!</b> "'.$selected['Title'].'"');
+		}
+		
 		$form = new form(
 				($edit?
 					__("Edit Ad / Banner"):
@@ -861,19 +879,11 @@ class _ads {
 				str_replace('&amp;', '&', url::uri('id, edit, delete'))."'\"");
 		}
 		
-		$selected = null;
 		$verifyok = false;
 		
-		if ($id)
-			$selected = sql::fetch(sql::run(
-				" SELECT `ID` FROM `{ads}`" .
-				" WHERE `ID` = '".$id."'" .
-				($this->userPermissionIDs?
-					" AND `ID` IN (".$this->userPermissionIDs.")":
-					null)));
-		
 		if ($this->userPermissionType & USER_PERMISSION_TYPE_WRITE &&
-			((!$edit && !$delete) || $selected))
+			((!$edit && !$delete) || !$this->userPermissionIDs ||
+			in_array($id, explode(',', $this->userPermissionIDs))))
 			$verifyok = $this->verifyAdmin($form);
 		
 		$rows = sql::run(
@@ -892,9 +902,10 @@ class _ads {
 					TOOLTIP_NOTIFICATION);
 		
 		if ($this->userPermissionType & USER_PERMISSION_TYPE_WRITE &&
-			(!$this->userPermissionIDs || ($edit && $selected)))
+			(!$this->userPermissionIDs || 
+			($edit && in_array($id, explode(',', $this->userPermissionIDs)))))
 		{
-			if ($edit && $selected && ($verifyok || !$form->submitted())) {
+			if ($edit && ($verifyok || !$form->submitted())) {
 				$selected = sql::fetch(sql::run(
 					" SELECT * FROM `{ads}`" .
 					" WHERE `ID` = '".$id."'"));
